JNG-6322 Separate Epsion from expression validation
