+++ adapter-psm/src/main/java/hu/blackbelt/judo/meta/expression/adapters/psm/ExpressionValidatorOnPsm.java
@@ -20,45 +20,28 @@ package hu.blackbelt.judo.meta.expression.adapters.psm;
  * #L%
  */
 
-import hu.blackbelt.epsilon.runtime.execution.ExecutionContext;
+import hu.blackbelt.judo.meta.expression.runtime.ExpressionValidationException;
 import org.slf4j.Logger;
-import hu.blackbelt.epsilon.runtime.execution.exceptions.ScriptExecutionException;
-import hu.blackbelt.judo.meta.expression.runtime.ExpressionEpsilonValidator;
-import hu.blackbelt.judo.meta.expression.runtime.ExpressionEvaluator;
 import hu.blackbelt.judo.meta.expression.runtime.ExpressionModel;
 import hu.blackbelt.judo.meta.psm.runtime.PsmModel;
-import org.eclipse.epsilon.common.util.UriUtil;
 
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
-import static hu.blackbelt.epsilon.runtime.execution.ExecutionContext.executionContextBuilder;
-import static hu.blackbelt.epsilon.runtime.execution.contexts.EvlExecutionContext.evlExecutionContextBuilder;
-import static hu.blackbelt.epsilon.runtime.execution.model.emf.WrappedEmfModelContext.wrappedEmfModelContextBuilder;
 import static java.util.Collections.emptyList;
 
-public class ExpressionEpsilonValidatorOnPsm extends ExpressionEpsilonValidator {
+public class ExpressionValidatorOnPsm {
 
-    public static void validateExpressionOnPsm(Logger log, PsmModel psmModel, ExpressionModel expressionModel, URI scriptRoot)
-            throws ScriptExecutionException, URISyntaxException {
-        validateExpressionOnPsm(log, psmModel, expressionModel, scriptRoot, emptyList(), emptyList());
+    public static void validateExpressionOnPsm(Logger log, PsmModel psmModel, ExpressionModel expressionModel)
+            throws ExpressionValidationException {
+        validateExpressionOnPsm(log, psmModel, expressionModel, emptyList(), emptyList());
     }
 
-    public static void validateExpressionOnPsm(Logger log, PsmModel psmModel, ExpressionModel expressionModel, URI scriptRoot,
+    public static void validateExpressionOnPsm(Logger log, PsmModel psmModel, ExpressionModel expressionModel,
                                                Collection<String> expectedErrors, Collection<String> expectedWarnings)
-            throws ScriptExecutionException, URISyntaxException {
-        validateExpressionOnPsm(log, psmModel, expressionModel,scriptRoot, expectedErrors, expectedWarnings, true);
+            throws ExpressionValidationException {
 
-    }
-
-    public static void validateExpressionOnPsm(Logger log, PsmModel psmModel, ExpressionModel expressionModel, URI scriptRoot,
-                                               Collection<String> expectedErrors, Collection<String> expectedWarnings, Boolean useCache)
-            throws ScriptExecutionException, URISyntaxException {
-
+        /*
         final Map<String, Object> injections = new HashMap<>();
         injections.put("evaluator", new ExpressionEvaluator());
         injections.put("modelAdapter", new PsmModelAdapter(psmModel.getResourceSet(), psmModel.getResourceSet()));
@@ -110,5 +93,6 @@ public class ExpressionEpsilonValidatorOnPsm extends ExpressionEpsilonValidator 
             } catch (Exception e) {
             }
         }
+        */
     }
 }
