+++ pom.xml
@@ -95,10 +95,10 @@
         <cxf-version>3.4.4</cxf-version>
         <pax-exam-version>4.13.5</pax-exam-version>
 
-        <judo-meta-expression-version>1.0.5.20250806_045448_5bfd94d3_develop</judo-meta-expression-version>
-        <judo-meta-psm-version>1.3.0.20250806_044805_33ba0f00_develop</judo-meta-psm-version>
-        <judo-meta-jql-version>1.0.4.20250806_043432_1336d9be_develop</judo-meta-jql-version>
-        <judo-meta-measure-version>1.0.2.20250806_044838_67c9254b_develop</judo-meta-measure-version>
+        <judo-meta-expression-version>1.0.5-SNAPSHOT</judo-meta-expression-version>
+        <judo-meta-psm-version>1.3.0-SNAPSHOT</judo-meta-psm-version>
+        <judo-meta-jql-version>1.0.4-SNAPSHOT</judo-meta-jql-version>
+        <judo-meta-measure-version>1.0.2-SNAPSHOT</judo-meta-measure-version>
 
         <!-- Define overridable properties for tycho-surefire-plugin -->
         <platformSystemProperties />
